% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphadiv.R
\name{alphadiv}
\alias{alphadiv}
\title{Basic alpha diversity calculation}
\usage{
alphadiv(comm, meta, group, cols = NULL, filename = deparse(substitute(comm)))
}
\arguments{
\item{comm}{Community matric, site (rows) by species (columns).}

\item{meta}{Metadata file; should have a column called 'SampleID' that matches the
Row names in \code{comm}.}

\item{group}{Grouping variable from metadata; name as character; May also indicate
a column for a numeric independent variable, rather than a group.}

\item{cols}{Character vector of length matching \code{group} levels; if NULL,
will default to color ramp of rainbow colors.}
}
\value{
Data frame containing alpha diversity metrics for each sample
and group assignment from metadata.
}
\description{
Use input community matrix and metadata to calculate richness, evenness,
Shannon diversity, and Chao1 indices and plot them with desired grouping variable.
}
\author{
Kara J.M. Taylor
}
