% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order3_interactions.R
\name{as.bipartite.net}
\alias{as.bipartite.net}
\title{Look at shared taxa between samples of different types}
\usage{
as.bipartite.net(X, Y, rare = F)
}
\arguments{
\item{X}{Input community matrix for first sample type}

\item{Y}{Input community matrix for second sample type}

\item{rare}{Logical or numeric; Rarefy the data first? If numeric, the number to rarefy to.}

\item{commList}{Output from \code{as.bipartite.net}}

\item{method}{Distance metric; See \code{vegan::vegdist} for options.}

\item{binary}{Logical; Whether the metric should use the binary version or not.}
}
\value{
A distance matrix formatted for downstream analysis.
}
\description{
Basically a specific format for a distance matrix between two input mattrices
}
\author{
Kara J.M. Taylor
}
