% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biom_as_csv.R
\name{biom.as.csv}
\alias{biom.as.csv}
\title{BIOM to CSV}
\usage{
biom.as.csv(path)
}
\arguments{
\item{path}{location of BIOM table}
}
\value{
Two .csv documents: one containing the community matrix (feature_table.csv)
and the other containing the taxonomy information (taxonomy.csv).
}
\description{
Converts *.biom files to CSV and extracts taxonomy info
}
\author{
Kara J.M. Taylor
}
