% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_regression.R
\name{compare_regressions}
\alias{compare_regressions}
\title{Compare regression coefficients}
\usage{
compare_regressions(m1, m2, coef)
}
\arguments{
\item{m1}{model 1}

\item{m2}{model 2}

\item{coef}{Either 'slope' or 'intercept'}
}
\description{
Use manual T-test calculation to test differences in regression coefficients
}
\author{
Kara J.M. Taylor
}
