% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendroheatmap_v2.R
\name{dendro.heatmap}
\alias{dendro.heatmap}
\title{Dendro-heatmap}
\usage{
dendro.heatmap(
  comm,
  tax,
  meta,
  bdiv = NULL,
  path = "",
  group,
  subgroup,
  core.list = 10,
  hi.tax = NULL
)
}
\arguments{
\item{comm}{Community matrix (site rows and species columns)}

\item{tax}{Taxonomy table}

\item{meta}{Sample metadata}

\item{bdiv}{Sample dissimilarity matrix (optional; recommend using UniFrac distance)}

\item{path}{Path to phylogenetic tree in Newick format.}

\item{group}{Column in \code{meta} indicating the sets}

\item{subgroup}{Column in \code{meta} indicating subsets or sample ID names}

\item{core.list}{List of ASVs to include. May be numeric to select the top n most abundant ASVs. Default= top 10.}

\item{hi.tax}{Taxonomic level in \code{tax} to summarize \code{core.list} to (optional)}

\item{tax.grp}{Taxonomic level to summarize taxa to. Defaults to 'tag'}
}
\value{
List containing set information
}
\description{
Constructs a sample abundance heatmap of select taxa with a
phylogenetic tree of taxa on the y-axis and a dendrogram of samples ordered
by hierarchical compositional relatedness on the x-axis.
}
\author{
Kara J.M. Taylor
}
