% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbiom_bar.R
\name{mbiom.bar}
\alias{mbiom.bar}
\title{Microbiome Venn diagram contents}
\usage{
mbiom.bar(
  shared,
  comm,
  select = NULL,
  tax,
  hi.tax,
  meta,
  group,
  subgroup,
  file = "mbiom_bar",
  print = F,
  compare.otus = F,
  set.y = F
)
}
\arguments{
\item{shared}{Output list from \code{shared_taxa}}

\item{comm}{Community matrix}

\item{select}{List of taxa to show; will show all taxa by default.}

\item{tax}{Parsed taxonomy table.}

\item{hi.tax}{Taxonomic level in \code{tax} to show in bar plots.}

\item{meta}{Metadata table}

\item{group}{Column in \code{meta} containing set grouping information}

\item{subgroup}{Column in \code{meta} containing within-set subdivisions of the taxonomic distribution.}

\item{file}{Output filename}

\item{print}{Logical; Print all plots to file? Default is FALSE}

\item{compare.otus}{Logical; Print plot showing summary distribution? Default is FALSE}

\item{set.y}{Logical; Fix ylim across all output plots? Default is FALSE}
}
\value{
Data frame containing relative abundance for each \code{hi.tax} in each group.subgroup level
}
\description{
Constructs bar graphs showing the relative abundance of taxa in each set.
}
\author{
Kara J.M. Taylor
}
