% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbiom_venn.R
\name{mbiom.venn}
\alias{mbiom.venn}
\title{Microbiome Venn diagram}
\usage{
mbiom.venn(mat, meta, group, tax = NULL, file = "", tax.grp = "tag")
}
\arguments{
\item{mat}{Community matrix (site rows and species columns)}

\item{meta}{Sample metadata}

\item{group}{Column in \code{meta} indicating the sets}

\item{tax}{Taxonomy table (optional)}

\item{file}{Output file name}

\item{tax.grp}{Taxonomic level to summarize taxa to. Defaults to 'tag'}
}
\value{
List containing set information
}
\description{
Constructs a venn diagram with the number of sets defined by levels in \code{group}.
For datasets where the number of groups is 5 or less, a diagram figure will be drawn.
For datasets with more than 5 groups, no figure will be drawn, but the sets will be listed.
}
\author{
Kara J.M. Taylor
}
