% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_taxa.R
\name{shared.taxa}
\alias{shared.taxa}
\title{Parsing shared taxa from Venn diagram}
\usage{
shared.taxa(mbiom, file, excel = T)
}
\arguments{
\item{mbiom}{Output list from \code{mbiom_venn}}

\item{file}{Output filename}

\item{excel}{Logical; return as .xlsx as well as to GlobalEnv?}
}
\value{
Returns list for each discrete segment in the Venn diagram
}
\description{
Assigns each taxon to its Venn digram subsection.
}
\author{
Kara J.M. Taylor
}
